Signed all engine libraries to allow for sandboxing. On the client: chat doesn't interfere with other commands; added range indicators; added a shader for simple shadowing Engine and Editor now share same object creation mechanism. Introduced a bug with editor where removing objects doesnt actually remove em. Same with engine.

+++ ShanoEditor/ViewModels/Scenario/ContentLists/Textures.cs
@@ -123,7 +123,8 @@ namespace ShanoEditor.ViewModels
             var fileName = Path.GetFileName(srcPath);
             var destFile = Path.Combine(destFolder, fileName);
 
-            File.Copy(srcPath, destFile);
+            if(!File.Exists(destFile))
+                File.Copy(srcPath, destFile);
         }
 
         public void MoveTexture(TextureViewModel tex, string newDir)
